+++ client/src/components/home/Featured.css
@@ -18,6 +18,7 @@
 
 .featured-card {
 	display: flex;
+
 	align-items: flex-end;
 	width: 100%;
 }
@@ -28,36 +29,27 @@
 }
 
 .game-image {
-	max-width: 400px;
-	height: 300px;
+	max-width: 450px;
+	height: 350px;
 	border-radius: 10px;
 	border: 2px solid #a661d0;
 	cursor: pointer;
 	transition: transform 0.3s ease, opacity 0.3s ease;
-	margin-left: 5%;
-}
-
-.device-container {
-	gap: 25px;
-}
-
-.content-box p {
-	font-weight: 500;
-	margin-top: 3%;
-	margin-bottom: 0%;
+	margin-left: 10%;
 }
 
 .thumb-text {
-	margin-left: 20px;
+	margin-left: 10%;
 	height: 43vh;
 	display: flex;
 	flex-direction: column;
+	justify-content: space-between;
 	align-items: center;
 }
 
 .game-thumbnail {
 	margin-top: 10px;
-	width: 200px;
+	width: 300px;
 	height: auto;
 	border-radius: 5px;
 	border: 2px solid #a661d0;
@@ -107,7 +99,7 @@
 
 .arrow.left {
 	left: 30px;
-	background-image: url("../assets/images/button_icons/buttomfleche.png");
+	background-image: url("../../assets/images/button_icons/buttomfleche.png");
 	background-color: transparent;
 }
 
@@ -118,6 +110,6 @@
 
 .arrow.right {
 	right: 30px;
-	background-image: url("../assets/images/button_icons/buttomflecheright.png");
+	background-image: url("../../assets/images/button_icons/buttomflecheright.png");
 	background-color: transparent;
 }
