+++ client/src/components/home/Featured.tsx
@@ -1,34 +1,26 @@
 import { useEffect, useState } from "react";
-
-import { useAuth } from "../../context/UserContext";
-
-import FavoriteButton from "../buttons/FavoriteButton";
-import GameShelfButton from "../buttons/GameShelfButton";
-import InfosButton from "../buttons/InfosButton";
-import GameDevices from "../game/GameDevices";
-
 import type GameType from "../../interface/GameType";
-
 import "./featured.css";
+import DOMPurify from "dompurify";
+import parse from "html-react-parser";
+import GameDevices from "../game/GameDevices";
 
 const Featured = () => {
 	const [games, setGames] = useState<GameType[]>([]);
 	const [currentIndex, setCurrentIndex] = useState(0);
 	const [isPrimaryImage, setIsPrimaryImage] = useState(true);
-	const { user } = useAuth();
 
 	useEffect(() => {
 		const fetchGames = async () => {
 			try {
-				const response = await fetch("http://localhost:3310/api/games");
+				const response = await fetch(
+					"http://localhost:3310/api/games/featured/today",
+				);
 				if (!response.ok) {
 					throw new Error("Failed to fetch games");
 				}
-				const allGames = await response.json();
-				const filteredGames = allGames.filter((game: GameType) =>
-					[1, 16, 32, 66].includes(Number(game.id)),
-				);
-				setGames(filteredGames);
+				const featuredGames = await response.json();
+				setGames(featuredGames);
 			} catch (error) {
 				console.error("Error fetching games:", error);
 			}
@@ -36,6 +28,20 @@ const Featured = () => {
 
 		fetchGames();
 	}, []);
+
+	const shortenDescription = (game: GameType) => {
+		const shortDescription = parse(
+			DOMPurify.sanitize(
+				game.description.length <= 200
+					? game.description
+					: `${game.description.slice(0, game.description.slice(0, 200).lastIndexOf(" "))}...`,
+				{
+					USE_PROFILES: { html: true },
+				},
+			),
+		);
+		return shortDescription;
+	};
 
 	const handleNext = () => {
 		if (games.length > 0) {
@@ -64,7 +70,7 @@ const Featured = () => {
 	};
 
 	return (
-		<div className="content-box homepage-content-box" id="featured">
+		<div className="content-box" id="featured">
 			<h2 className="featured-title">Featured Games</h2>
 			<button
 				type="button"
@@ -74,91 +80,78 @@ const Featured = () => {
 			/>
 
 			{games.length > 0 && (
-				<>
-					<div className="featured-card">
+				<div className="featured-card">
+					<img
+						src={
+							isPrimaryImage
+								? games[currentIndex].image
+								: games[currentIndex].image_2
+						}
+						alt={games[currentIndex]?.title || "Game"}
+						className="game-image"
+						onClick={toggleImage}
+						onKeyDown={toggleImage}
+					/>
+					<div className="thumb-text">
+						<h3 className="game-name">{games[currentIndex]?.title}</h3>
+
+						<div className="device-container">
+							{Array.isArray(games[currentIndex]?.devices) ? (
+								<GameDevices devices={games[currentIndex].devices} />
+							) : typeof games[currentIndex]?.devices === "string" ? (
+								<GameDevices
+									devices={(games[currentIndex].devices as string)
+										.split(",")
+										.map(
+											(device: string) =>
+												device.trim() as
+													| "PlayStation"
+													| "Nintendo"
+													| "PC"
+													| "Xbox"
+													| "Others",
+										)}
+								/>
+							) : (
+								<p>Devices: Unknown</p>
+							)}
+						</div>
+
+						<p className="game-genres">
+							{Array.isArray(games[currentIndex]?.genres)
+								? games[currentIndex]?.genres.join(", ")
+								: games[currentIndex]?.genres || "Unknown"}
+						</p>
+
+						<p className="short-description">
+							{shortenDescription(games[currentIndex]) && (
+								<p>{shortenDescription(games[currentIndex])}</p>
+							)}
+						</p>
+
 						<img
 							src={
 								isPrimaryImage
-									? games[currentIndex].image
-									: games[currentIndex].image_2
+									? games[currentIndex].image_2
+									: games[currentIndex].image
 							}
-							alt={games[currentIndex]?.title || "Game"}
-							className="game-image"
-							onClick={toggleImage}
-							onKeyDown={toggleImage}
+							alt="Thumbnail"
+							className="game-thumbnail"
+							onClick={handleThumbnailClick}
+							onKeyDown={handleThumbnailClick}
 						/>
-						<div className="thumb-text">
-							<h3 className="game-name">{games[currentIndex]?.title}</h3>
-
-							<div className="device-container">
-								{Array.isArray(games[currentIndex]?.devices) ? (
-									<GameDevices devices={games[currentIndex].devices} />
-								) : typeof games[currentIndex]?.devices === "string" ? (
-									<GameDevices
-										devices={(games[currentIndex].devices as string)
-											.split(",")
-											.map(
-												(device: string) =>
-													device.trim() as
-														| "PlayStation"
-														| "Nintendo"
-														| "PC"
-														| "Xbox"
-														| "Others",
-											)}
-									/>
-								) : (
-									<p>Devices: Unknown</p>
-								)}
-							</div>
-
-							<p className="game-genres">
-								{" "}
-								{Array.isArray(games[currentIndex]?.genres)
-									? games[currentIndex]?.genres.join(", ")
-									: games[currentIndex]?.genres || "Unknown"}
-							</p>
-
-							<p className="game-tags">
-								#{" "}
-								{Array.isArray(games[currentIndex]?.tags)
-									? games[currentIndex]?.tags.join(", ")
-									: games[currentIndex]?.tags || "Unknown"}
-							</p>
-
-							<img
-								src={
-									isPrimaryImage
-										? games[currentIndex].image_2
-										: games[currentIndex].image
-								}
-								alt="Thumbnail"
-								className="game-thumbnail"
-								onClick={handleThumbnailClick}
-								onKeyDown={handleThumbnailClick}
-							/>
-						</div>
 					</div>
-
-					<div className="game-actions">
-						<InfosButton id={games[currentIndex].id} />
-						{user ? (
-							<>
-								<GameShelfButton
-									userId={user.id}
-									gameId={Number.parseInt(games[currentIndex].id)}
-								/>
-								<FavoriteButton
-									userId={user.id}
-									gameId={Number.parseInt(games[currentIndex].id)}
-								/>
-							</>
-						) : (
-							<></>
-						)}
-					</div>
-				</>
+				</div>
 			)}
+
+			<div className="game-actions">
+				<button type="button" className="beautiful-buttonadd">
+					i
+				</button>
+				<button type="button" className="beautiful-buttonadd">
+					+
+				</button>
+			</div>
 
 			<button
 				type="button"
